--- a/ecmascript/transforms/src/optimization/simplify/dce/mod.rs
+++ b/ecmascript/transforms/src/optimization/simplify/dce/mod.rs
@@ -5,7 +5,6 @@ use retain_mut::RetainMut;
 use std::{any::type_name, borrow::Cow, mem::take};
 use swc_atoms::JsWord;
 use swc_common::{
-    chain,
     pass::{CompilerPass, Repeated},
     util::move_map::MoveMap,
     BytePos, Mark, Span, Spanned,
@@ -17,7 +16,7 @@ use swc_ecma_visit::{as_folder, noop_visit_mut_type, VisitMut, VisitMutWith, Vis
 macro_rules! preserve {
     ($name:ident, $T:ty) => {
         fn $name(&mut self, node: &mut $T) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
         }
     };
 }
@@ -109,24 +108,24 @@ impl VisitMut for Dce<'_> {
     preserve!(visit_mut_continue_stmt, ContinueStmt);
 
     fn visit_mut_block_stmt(&mut self, node: &mut BlockStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
         node.stmts.visit_mut_with(self);
 
-        if self.marking_phase || node.stmts.iter().any(|stmt| self.is_marked(stmt.span())) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+        if self.marking_phase || node.stmts.iter().any(|stmt| self.is_used(stmt.span())) {
+            self.preserve(node.span);
             self.mark(&mut node.stmts);
         }
     }
 
     fn visit_mut_class_decl(&mut self, node: &mut ClassDecl) {
-        if self.is_marked(node.span()) {
+        if self.is_used(node.span()) {
             return;
         }
 
         if self.marking_phase || self.included.contains(&node.ident.to_id()) {
-            node.class.span = node.class.span.apply_mark(self.config.used_mark);
+            self.preserve(node.class.span);
             self.mark(&mut node.class.super_class);
         }
 
@@ -134,14 +133,14 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_do_while_stmt(&mut self, node: &mut DoWhileStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.visit_mut_children_with(self);
 
-        if self.is_marked(node.test.span()) || self.is_marked(node.body.span()) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+        if self.is_used(node.test.span()) || self.is_used(node.body.span()) {
+            self.preserve(node.span);
 
             self.mark(&mut node.test);
             self.mark(&mut node.body);
@@ -149,14 +148,14 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_export_all(&mut self, node: &mut ExportAll) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
-        node.span = node.span.apply_mark(self.config.used_mark);
+        self.preserve(node.span);
     }
 
     fn visit_mut_export_decl(&mut self, node: &mut ExportDecl) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
@@ -165,14 +164,14 @@ impl VisitMut for Dce<'_> {
 
             // Preserve types
             Decl::TsInterface(_) | Decl::TsTypeAlias(_) | Decl::TsEnum(_) | Decl::TsModule(_) => {
-                node.span = node.span.apply_mark(self.config.used_mark);
+                self.preserve(node.span);
                 return;
             }
 
             // Preserve only exported variables
             Decl::Var(v) => {
                 v.decls.retain_mut(|d| {
-                    if self.is_marked(d.span()) {
+                    if self.is_used(d.span()) {
                         return true;
                     }
 
@@ -181,7 +180,7 @@ impl VisitMut for Dce<'_> {
                         if self.included.iter().any(|included| *included == name)
                             || self.should_preserve_export(&name.0)
                         {
-                            d.span = d.span.apply_mark(self.config.used_mark);
+                            self.preserve(d.span);
                             self.mark(&mut d.init);
                             return true;
                         }
@@ -194,7 +193,7 @@ impl VisitMut for Dce<'_> {
                 });
 
                 if self.decl_dropping_phase && !v.decls.is_empty() {
-                    node.span = node.span.apply_mark(self.config.used_mark);
+                    self.preserve(node.span);
                 }
 
                 return;
@@ -202,42 +201,42 @@ impl VisitMut for Dce<'_> {
         };
 
         if self.should_preserve_export(&i.sym) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
             self.mark(&mut node.decl);
         }
     }
 
     fn visit_mut_export_default_decl(&mut self, node: &mut ExportDefaultDecl) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         // TODO: Export only when it's required. (i.e. check self.used_exports)
 
-        node.span = node.span.apply_mark(self.config.used_mark);
+        self.preserve(node.span);
         self.mark(&mut node.decl);
     }
 
     fn visit_mut_export_default_expr(&mut self, node: &mut ExportDefaultExpr) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         if self.should_preserve_export(&js_word!("default")) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
             self.mark(&mut node.expr);
         }
     }
 
     fn visit_mut_expr_stmt(&mut self, node: &mut ExprStmt) {
         log::debug!("ExprStmt ->");
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         if self.should_include(&node.expr) {
             log::debug!("\tIncluded");
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
             self.mark(&mut node.expr);
             return;
         }
@@ -245,13 +244,13 @@ impl VisitMut for Dce<'_> {
         node.visit_mut_children_with(self)
     }
 
-    fn visit_mut_fn_decl(&mut self, mut f: &mut FnDecl) {
-        if self.is_marked(f.span()) {
+    fn visit_mut_fn_decl(&mut self, f: &mut FnDecl) {
+        if self.is_used(f.span()) {
             return;
         }
 
         if self.marking_phase || self.included.contains(&f.ident.to_id()) {
-            f.function.span = f.function.span.apply_mark(self.config.used_mark);
+            self.preserve(f.function.span);
             self.mark(&mut f.function.params);
             self.mark(&mut f.function.body);
             return;
@@ -261,7 +260,7 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_for_in_stmt(&mut self, node: &mut ForInStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
@@ -269,11 +268,11 @@ impl VisitMut for Dce<'_> {
         node.body.visit_mut_with(self);
 
         if self.should_include(&node.left)
-            || self.is_marked(node.left.span())
-            || self.is_marked(node.right.span())
-            || self.is_marked(node.body.span())
+            || self.is_used(node.left.span())
+            || self.is_used(node.right.span())
+            || self.is_used(node.body.span())
         {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
 
             self.mark(&mut node.left);
             self.mark(&mut node.right);
@@ -282,7 +281,7 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_for_of_stmt(&mut self, node: &mut ForOfStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
@@ -290,11 +289,11 @@ impl VisitMut for Dce<'_> {
         node.body.visit_mut_with(self);
 
         if self.should_include(&node.left)
-            || self.is_marked(node.left.span())
-            || self.is_marked(node.right.span())
-            || self.is_marked(node.body.span())
+            || self.is_used(node.left.span())
+            || self.is_used(node.right.span())
+            || self.is_used(node.body.span())
         {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
 
             self.mark(&mut node.left);
             self.mark(&mut node.right);
@@ -303,19 +302,19 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_for_stmt(&mut self, node: &mut ForStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.visit_mut_children_with(self);
 
         if node.test.is_none()
-            || self.is_marked(node.init.span())
-            || self.is_marked(node.test.span())
-            || self.is_marked(node.update.span())
-            || self.is_marked(node.body.span())
+            || self.is_used(node.init.span())
+            || self.is_used(node.test.span())
+            || self.is_used(node.update.span())
+            || self.is_used(node.body.span())
         {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
 
             self.mark(&mut node.test);
             self.mark(&mut node.init);
@@ -325,7 +324,7 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_ident(&mut self, i: &mut Ident) {
-        if self.is_marked(i.span) {
+        if self.is_used(i.span) {
             return;
         }
 
@@ -338,17 +337,17 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_if_stmt(&mut self, node: &mut IfStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.visit_mut_children_with(self);
 
-        if self.is_marked(node.test.span())
-            || self.is_marked(node.cons.span())
-            || self.is_marked(node.alt.span())
+        if self.is_used(node.test.span())
+            || self.is_used(node.cons.span())
+            || self.is_used(node.alt.span())
         {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
 
             self.mark(&mut node.test);
             self.mark(&mut node.cons);
@@ -356,19 +355,19 @@ impl VisitMut for Dce<'_> {
         }
     }
 
-    fn visit_mut_import_decl(&mut self, mut import: &mut ImportDecl) {
+    fn visit_mut_import_decl(&mut self, import: &mut ImportDecl) {
         // Do not mark import as used while ignoring imports
         if !self.decl_dropping_phase {
             return;
         }
 
-        if self.is_marked(import.span) {
+        if self.is_used(import.span) {
             return;
         }
 
         // Side effect import
         if import.specifiers.is_empty() {
-            import.span = import.span.apply_mark(self.config.used_mark);
+            self.preserve(import.span);
             return;
         }
 
@@ -377,25 +376,25 @@ impl VisitMut for Dce<'_> {
         import.specifiers.retain(|s| self.should_include(s));
 
         if !import.specifiers.is_empty() {
-            import.span = import.span.apply_mark(self.config.used_mark);
+            self.preserve(import.span);
         }
     }
 
     fn visit_mut_labeled_stmt(&mut self, node: &mut LabeledStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.body.visit_mut_with(self);
 
-        if self.is_marked(node.body.span()) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+        if self.is_used(node.body.span()) {
+            self.preserve(node.span);
             self.mark(&mut node.body);
         }
     }
 
     fn visit_mut_member_expr(&mut self, e: &mut MemberExpr) {
-        if self.is_marked(e.span()) {
+        if self.is_used(e.span()) {
             return;
         }
 
@@ -406,7 +405,7 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_named_export(&mut self, node: &mut NamedExport) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
@@ -420,29 +419,29 @@ impl VisitMut for Dce<'_> {
         });
 
         if !node.specifiers.is_empty() {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
             self.mark(&mut node.specifiers);
         }
     }
 
     fn visit_mut_return_stmt(&mut self, node: &mut ReturnStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
-        node.span = node.span.apply_mark(self.config.used_mark);
+        self.preserve(node.span);
         self.mark(&mut node.arg);
     }
 
     fn visit_mut_switch_case(&mut self, node: &mut SwitchCase) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.visit_mut_children_with(self);
 
-        if self.is_marked(node.test.span()) || node.cons.iter().any(|v| self.is_marked(v.span())) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+        if self.is_used(node.test.span()) || node.cons.iter().any(|v| self.is_used(v.span())) {
+            self.preserve(node.span);
 
             self.mark(&mut node.test);
             self.mark(&mut node.cons);
@@ -450,7 +449,7 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_switch_stmt(&mut self, node: &mut SwitchStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
@@ -462,39 +461,39 @@ impl VisitMut for Dce<'_> {
         //            self.is_marked(case.span)
         //        });
 
-        if self.is_marked(node.discriminant.span())
-            || node.cases.iter().any(|case| self.is_marked(case.span))
+        if self.is_used(node.discriminant.span())
+            || node.cases.iter().any(|case| self.is_used(case.span))
         {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
             self.mark(&mut node.cases);
         }
     }
 
     fn visit_mut_throw_stmt(&mut self, node: &mut ThrowStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
-        node.span = node.span.apply_mark(self.config.used_mark);
+        self.preserve(node.span);
 
         node.visit_mut_children_with(self);
 
-        if self.is_marked(node.arg.span()) {
+        if self.is_used(node.arg.span()) {
             self.mark(&mut node.arg)
         }
     }
 
     fn visit_mut_try_stmt(&mut self, node: &mut TryStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.visit_mut_children_with(self);
 
-        if self.is_marked(node.block.span())
-            || self.is_marked(node.handler.span())
-            || self.is_marked(node.finalizer.span())
+        if self.is_used(node.block.span())
+            || self.is_used(node.handler.span())
+            || self.is_used(node.finalizer.span())
         {
-            node.span = node.span.apply_mark(self.config.used_mark);
+            self.preserve(node.span);
 
             self.mark(&mut node.block);
             self.mark(&mut node.handler);
@@ -503,23 +502,23 @@ impl VisitMut for Dce<'_> {
     }
 
     fn visit_mut_update_expr(&mut self, node: &mut UpdateExpr) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
-        node.span = node.span.apply_mark(self.config.used_mark);
+        self.preserve(node.span);
         self.mark(&mut node.arg);
     }
 
-    fn visit_mut_var_decl(&mut self, mut var: &mut VarDecl) {
-        if self.is_marked(var.span) {
+    fn visit_mut_var_decl(&mut self, var: &mut VarDecl) {
+        if self.is_used(var.span) {
             return;
         }
 
         var.visit_mut_children_with(self);
 
         var.decls.retain_mut(|decl| {
-            if self.is_marked(decl.span()) {
+            if self.is_used(decl.span()) {
                 return true;
             }
 
@@ -530,7 +529,7 @@ impl VisitMut for Dce<'_> {
                 return true;
             }
 
-            decl.span = decl.span.apply_mark(self.config.used_mark);
+            self.preserve(decl.span);
             self.mark(&mut decl.init);
             self.mark(&mut decl.name);
             true
@@ -540,18 +539,18 @@ impl VisitMut for Dce<'_> {
             return;
         }
 
-        var.span = var.span.apply_mark(self.config.used_mark);
+        self.preserve(var.span);
     }
 
     fn visit_mut_while_stmt(&mut self, node: &mut WhileStmt) {
-        if self.is_marked(node.span) {
+        if self.is_used(node.span) {
             return;
         }
 
         node.visit_mut_children_with(self);
 
-        if self.is_marked(node.test.span()) || self.is_marked(node.body.span()) {
-            node.span = node.span.apply_mark(self.config.used_mark);
+        if self.is_used(node.test.span()) || self.is_used(node.body.span()) {
+            self.preserve(node.span);
 
             self.mark(&mut node.test);
             self.mark(&mut node.body);
@@ -636,7 +635,7 @@ impl Dce<'_> {
 
                 idx += 1;
                 // Drop unused imports
-                if self.is_marked(item.span()) {
+                if self.is_used(item.span()) {
                     Some(item)
                 } else {
                     None
@@ -649,20 +648,12 @@ impl Dce<'_> {
 }
 
 impl Dce<'_> {
-    pub fn is_marked(&mut self, span: Span) -> bool {
-        let mut ctxt = span.ctxt().clone();
+    fn preserve(&mut self, span: Span) {
+        self.preserved.insert((span.lo, span.hi));
+    }
 
-        loop {
-            let mark = ctxt.remove_mark();
-
-            if mark == Mark::root() {
-                return false;
-            }
-
-            if mark == self.config.used_mark {
-                return true;
-            }
-        }
+    fn is_used(&mut self, span: Span) -> bool {
+        self.preserved.contains(&(span.lo, span.hi))
     }
 
     pub fn should_preserve_export(&self, i: &JsWord) -> bool {