## dce

### Baseline

```
test dce_16 ... bench:       9,026 ns/iter (+/- 901) = 3 MB/s
test dce_4  ... bench:       2,778 ns/iter (+/- 442) = 3 MB/s
```

### HashSet

```
test dce_16 ... bench:       8,062 ns/iter (+/- 616) = 4 MB/s
test dce_4  ... bench:       2,547 ns/iter (+/- 217) = 4 MB/s
```